fix : collection files, import in about and gitkeep
