+++ package.json
@@ -24,6 +24,7 @@
     "lucide-react": "0.427.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "sharp": "0.33.5",
     "tailwind-merge": "2.4.0",
     "tailwindcss": "^3.4.9",
     "tailwindcss-animate": "1.0.7",
