+++ src/content/events/rouen-france-2024.mdx
@@ -214,6 +214,6 @@ content:
     href: https://www.youtube.com/watch?v=HLGnkpZ-Q1Q
 ---
 
-import { About } from "./2024-06-07-rouen/about";
+import { About } from "./rouen-france-2024/about";
 
 <About event={frontmatter} />
